feat: update server file structure
